Fix dead code: remove unused HeadRoute and restore @glint-nocheck for ember-cli-head template
- Delete app/routes/head.ts which was dead code (no 'head' route exists in router)
- Restore @glint-nocheck directive for head.hbs since it's a special ember-cli-head
  template that cannot be typed via a route
- Add HeadContent to Glint registry to provide type safety for head-layout.hbs
