+++ types/glint.d.ts
@@ -5,6 +5,20 @@ import type EmberConcurrencyRegistry from 'ember-concurrency/template-registry';
 import type EmberTruthHelpersRegistry from 'ember-truth-helpers/template-registry';
 import type EmberMathHelpersRegistry from 'ember-math-helpers/template-registry';
 import type { ModifierLike, HelperLike, ComponentLike } from '@glint/template';
+import type HeadDataService from 'codecrafters-frontend/services/meta-data';
+
+interface HeadTemplateModel {
+  metaData: HeadDataService;
+  defaults: {
+    description: string;
+    type: string;
+    siteName: string;
+    title: string;
+    imageUrl: string;
+    twitterCard: string;
+    twitterSite: string;
+  };
+}
 
 declare module '@glint/environment-ember-loose/registry' {
   export default interface Registry
@@ -44,6 +58,9 @@ declare module '@glint/environment-ember-loose/registry' {
 
     'sortable-handle': ModifierLike<{ Args: { Positional: [] } }>;
     'svg-jar': ComponentLike<{ Args: { Named: { class: string; role?: string }; Positional: [string] } }>;
+
+    // ember-cli-head: HeadContent component renders app/templates/head.hbs with @model
+    HeadContent: ComponentLike<{ Args: { Named: { model: HeadTemplateModel } } }>;
     // ...
   }
 }
